Added preliminary speaker bio functionality
Note that this is still not fully functional. The issue is that the directory containing the bios has to be outside of the events data folder, or else the thing that looks through it tries to parse the ones below.

There is also an issue where the folder cannot have multiple trees - the space needed in the query is not coming through in the `printf` command.

For now this is a start.


Former-commit-id: 43dc55e6faa1b75498228725823eb604afc245ca