Add 2012 italy proposals via static
This works in theory - what I did was copy the output files in the the `/static/events/2012-italy/` folder.

The problem is the sponsor bar is messed up. I don't think we want to copy the files over though - it's just ugly and adds cruft. But at least this way the content doesn't break?


Former-commit-id: 58ddc36020a86529588d01b4346e8356af892aff