Refactor future footer
It's somewhat sloppy, and has the year 2016 hardcoded into it for now, but it works. This removes the issue with the "events on the same date" (only for future footer; the issue still exists for past events for now) and it also takes care of the "year in quotes" issue.

Resolves #247
Resolves #246
Resolves #233