[DC 2016] Closing out DC 2016
This advice in content/data/2016-washington-dc.yml appears to be
incorrect with regard to the quotes:

    year: "2016" # The year of the event. Make sure it is in quotes.

With quotes there, the event does not appear in the "past" partial.
Without quotes, it does. Thus I've removed the quotes, but feel free to
tell me how to fix this another way if that breaks anything.