(GH-641) Add Wercker step to check for invalid Windows filenames
This commit adds a helper script (build/check-invalid-filenames.sh)
which processes the file and directory names in the repoistory and
reports an error if it finds names which would be invalid on a
Windows filesystem.

This commit configures Wercker to run the validation script prior to
building via Hugo.  The script will fail on the first file it finds.
To list all invalid files, set environment variable NO_EXIT_ON_FAILURE
and the script will list all errors on STDERR but emit a zero exit code.