(GH989) Fix building Hugo website on Windows (#1085)
* (GH989) Fix building Hugo website on Windows

fixes #989

Previously the `File.Path` property would use OS dependant path seperator, back
slash on Windows and forward slash on *nix and Mac.  However as we are parsing
the event slug based on a path splitting this is problematic.

This commit changes the path splitting character to a Site level parameter which
can be changed prior to building via Hugo. The default value should be set to
`/` otherwise all of the building automation will fail, amongst other things.

A later commit will add documentation around this configuration item.

* (GH989) Add appveyor PR configuration file

This commit adds appveyor.yml which is used to build the devopsdays website
using Hugo on a Windows platform.  This is intended to be used to verify Pull
Requests prior to merging.