[LDN-2018] We have a new logo!
We felt the old logo wasn't really good enough so we went and got a new
one. The old one was there since 2016 and the major change to it was to
make it purple. This one is way nicer and comes with various elements we
can add to our marketing.

Signed-off-by: Chris M <millscj01@gmail.com>

[LDN-2018] Reduce size of new logo/optimise

Optimise images so they're not massive.

Signed-off-by: Chris M <millscj01@gmail.com>