Add basic sponsorship packages and welcome.
Start by adding the basic sponsorship packages and a bunch of clickable
stuff on the main welcome page. We've not published a prospectus yet but
we have a date.

---

Update with stand sizes and number of tickets

I've just found out the number of tickets we get allocated, so I'll add
these onto this PR.

Signed-off-by: Chris Mills <me@christophermills.co.uk>