[LDN-2019] Enable accessibility and input CFP dates. (#5957)
* Enable accessibility and input CFP dates.

We've got some CFP dates now so I've added those but not enabled the
page as of yet. I've also enabled the accessibility page as it's
worthwhile having this after discussing in our weekly meeting.

---

We provide a quiet room

Signed-off-by: Chris Mills <me@christophermills.co.uk>

* Comment out dates for time being

If CFP dates are published it's displayed. We shouldn't show that for
the time being.

Signed-off-by: Chris Mills <me@christophermills.co.uk>