add new rule around exchange opportunities
We all agree that it's important for devopsdays conferences to actively promote opportunities for exchange and sharing (see: CAMS); however, opinions are mixed on open spaces _specifically_, both at the event level and within the core as well. I'd like to add language to mandate the ideal that we agree on, while leaving it sufficiently open to local interpretation / execution.