Fix grammar for becoming a sponsor. (#7468)
Boston recently added an a-la-carte sponsorship type. The framework
duly creates a grammatically incorrect "Become a a la carte sponsor"
link on our page.

This change makes vowels form grammatically correct sentences.
The solution is not perfect - it will screw up on things like
"Become an one-day sponsor," but is generally more often correct
than not.