Update 2020-krakow.yml
We don't have an email address for Jeroen Fossaert so I can't add them - please email info@devopsdays.org when you have their email, and submit a PR to re-add them.

Also, I removed the dates because this PR did not include an address. We require a specific confirmed location if you're going to publish dates.