Let SELinux-protected hosts mount in site data. (#9699)
This is a change I've been carrying forward for some time now in my local clone.

The `cached` option is specific to Docker for Mac, and has no bearing on a Linux
host. Linux hosts with SELinux enabled will end up with a container that can't
read the local copy of the site data, unless Docker is told to change labels on
the files beforehand.

**I did not test this on a Mac.** That is why I have the global `else` that makes
any non-Linux host behave the old way.