Automate the process of changing a sponsor logo.
This script implements the process defined in the [instructions for updating a sponsor image](https://github.com/devopsdays/devopsdays-web/blob/master/utilities/README.md#updating-a-sponsor-image-or-data). I also library-ified `sedcmd`, as I wanted it in my script, but didn't want to copy-paste it.

The output of this script can be seen at #9903.

**Note that I only tested this on a Linux box.** I don't know that I haven't broken any functionality on macOS.