[buf-2020] Correct Sponsor logo for Stark & Wayne
Per Sponsor request, we should always be using the vertical blue logo,
with the shield iconography above the name.  I applied this to 2019 as
well, since doing this for 2020 resulted in the same image twice, with
the exact same checksums, and I felt this was cleaner.

This also reverts commit 1b0a118b1bbc2666876ce3dd09ac66a30394b90e.