[BOS-2021] Change a-la-carte sponsorship tiers
We are updating our a-la-carte sponsorship tiers to be more amenable
to the needs of a fully virtual conference, where sponsors cannot
justify a $5k or $8k price tag without screen time. We are also
dropping the pre-event hangout entirely, as it is not feasible to run
virtually with a sponsor, plus the videography, which is a built-in
of the virtual platform.