Sponsor levels with `max: 1` are suffixed with plurals
I noticed when working on #11449 that sponsors with the `max: 1` flag
set still are described as "Sponsors", rather than "Sponsor". The
comparison is for the string `"1"` rather than the integer `1`. I chose
to change the comparison to `1` and update events that had used `"1"`,
because there are many more events using `1` than `"1"` today.