[BOS-2023] Launch the Boston 2023 event
We nailed down our date and venue for this year's event, and would
like to launch a skeleton site as a placeholder. We are using last
year's logo and copy for now.

Aaron, Derek, and Jirawat have stepped back fron organizing this
year - I'll mail the global DL to confirm.