[LDN-2023] Add John Hatton Picture, coffee times and sponsor change
I've added a picture for John hatton

Added the times we're going to serve coffee and other breaks.

Removed devops exhange added in my previous commit (not used anywhere
else) and moved it to doxlon.

They're the same company but this is the one they want to use but with
a soon to be updated image and url etc which I'll update once we get
more information from them and backdate the previous years where they've
been present (2018 I think)