[ORG-2023] Update schedule
Reflects the starting time of the social event.
Moves introduction of the Open Spaces process to the Introduction.
I reviewed the time we have on the schedule to switch from session to session, which seems like enough buffer time. Might need to review this if @jjasghar has wild plans for virtual.