Move archived assets to CDN
This change imports assets from archived events (2015-2018) into the CDN. Since they are no longer actively maintained or consumed by Hugo templating, they are not needed in the main site.

Many of these assets (e.g. all of the `@2x.png` files) do not appear to be used on any pages, but I have not made any attempt to remove unreferenced assets.

I've included the script which did this, at `utilities/archive_assets.sh`. Note that I only tested this script under Linux, not Windows or macOS.

There were only two name clashes between the CDN - `sponsorship-packages.png` and `sponsorship-exclusive-packages.png` in the Oslo 2017 event. These appear to be unreferenced, so I've removed them from devopsdays-web, and made no change to the existing ones in devopsdays-assets. The script detects name clashes and forcefully exits, requiring the user to resolve a name clash before trying again.

The script is slower than it could be (~1-2min per event year), because it's n^2 over the number of assets for a given year (it for-each'es over assets in the event, _then_ the event's HTML files for each one). I could fix this, but it didn't seem worth it. I'm happy to fix that if anybody disagrees.

I have only spot-checked some HTML files with `git diff`. I'd like to merge the paired PR in devopsdays-assets, run the site locally off this branch, and spot-check events again, before I give this PR a clean bill of health.