[LDN-2023] Add Gitea and Elastic as sponsors
I've added two sponsors with Gitea not existing.

We didn't get much indication on a logo from them so I took the one from
wikimedia/the svg on their site and converted it into png.

As for the URL they 301 to about.gitea.com so it felt logical to put
that as their URL in data.

I can alwways change this once we hear back from them.