Updates the Good, Better Best blog
Adds some missing words to a sentence.

From:
But often the total number of attendees and how the event actually felt on the ground, including attendees, sponsors, and organizers.

To:
But don’t often have visibility into the total number of attendees and how the event actually felt on the ground, including attendees, sponsors, and organizers.

Signed-off-by: Nathen Harvey <nathenharvey@google.com>